create module.js, fix eslint and formatting
