+++ .eslintrc.js
@@ -1,10 +1,19 @@
 module.exports = {
   env: {
+    browser: true,
     es6: true,
   },
   extends: [
     'airbnb-base',
   ],
+  globals: {
+    Atomics: 'readonly',
+    SharedArrayBuffer: 'readonly',
+  },
+  parserOptions: {
+    ecmaVersion: 2018,
+    sourceType: 'module',
+  },
   rules: {
     'no-console': 0,
   },
