+++ package.json
@@ -18,6 +18,7 @@
   "devDependencies": {
     "eslint": "^6.8.0",
     "eslint-config-airbnb-base": "^14.0.0",
+    "eslint-plugin-import": "^2.20.0",
     "nodemon": "^2.0.2"
   }
 }
